fix(vbe): change framebuffer_addr type to uint64_t and make vbe_info static

+++ drivers/display/vbe/vbe.h
@@ -13,7 +13,7 @@ enum vbe_result {
 };
 
 typedef struct s_vbe_info {
-    uint32_t framebuffer_addr;
+    uint64_t framebuffer_addr;
     uint32_t width;
     uint32_t height;
     uint32_t pitch;
